requirements.txt
openai==1.65.1
python-dotenv==1.0.1
rispy==0.9.0
